Update manifest to use new buildpack
Previously the manifest for this application was configured such that
it used a custom command to start itself (based on Spring Boot).  This
change removes that custom command and substitutes in the new Java
buildpack which takes care of all of the things that command
previously did.  A byproduct of this is that as improvements are made
to the main buildpack they will automatically get included in future
pushes of the application.  An example of this is that now, if a New
Relic service is bound to the application, the application will
automatically be configured to use it.

[#54518830]