Fix maven warning
This commit fixes a warning displayed by Maven due to the fact that
no relativePath is set and maven defaults to the parent directory. As
the parent is never going to be available in such directory and we use
the maven repository to fetch it, this commit sets an unknown relative
path explicitly.

Also fixing to UTF-8 encoding which is a sane default. Fixing the
encoding removes some warning when copying filtered resources.