Update README with instructions for deploying app
The instructions should be explicit and precise enough for
anyone to deploy to production now. There is a Bamboo job
that should be doing it, but we've had some issue with it and
it often fails, whereas manual pushes always work. You need
a Cloudfoundry account with access to the Sagan organization
to push to production (ask Trevor Marshall).