Restore the default starter if none is set
This commit fixes a regression introduced by the library refactoring.

Previously, if no dependency was selected, the default
'spring-boot-starter' was added to provide the necessary base
dependencies.

This commit adds a addDefaultDependency on ProjectRequest that
adds that dependency. It can be overridden if the default needs to be
different.

Fixes gh-34