Prevent invalid application name
Previously, if a user choose 'SpringBoot' or 'Spring' as the name of the
project, the service will generate a `SpringBootApplication` or
`SpringApplication` respectively. Both of which leads to a compilation
failure since those names are already used in the current context.

The generation of the application name based on the project's name have
been moved to InitializrMetadata and two new properties have been
introduced:

* env.fallbackApplicationName defines the name of the application if the
one that was generated was invalid for some reasons
* env.invalidApplicationNames defines a list of application names that
should flagged as invalid. When the current candidate is equal to one of
them, the fallback should be used instead

These properties have default values that prevent such issue to happen
by default.

Fixes gh-79