Better describe service capability
Make sure that the title and description of each capability is
configurable and expose that information in the meta-data.

This typically allows to customize the header of each capability in the
UI and the description in help usage such as the optimized text output
for command line clients.

See gh-89