Polish template
I like Spring Initializr (numerous references on twitter) and it feels
odd to me to put the address of the web site on the web site itself since
we want to push for custom deployments (that would have a different URL
obviously).

Also the "Quick add dependencies" can be simply named "Dependencies"; it
is consistent with the project metadata section as well.