Additional dependency meta-data
This commit add two additional attributes to a dependency: weight
and keywords.

Weight is a numeric value (default to 0) that allows to order
dependencies search result in a sensible manner. This is a UI feature
only and is not exposed in the meta-data.

Keywords provides a list of words that are added to the source of terms
that the UI use to discover dependencies. It's not exposed either.

Both are exposed by the UI specific controller (`/ui/dependencies`).

See gh-145