Expose dependencies meta-data
Register an additional endpoint at `/dependencies` that returns the
necessary information about the registered dependencies for a Spring
Boot version. If the `bootVersion` request parameter is not specified,
the default Spring Boot version is used.

The json content contains the artifact coordinates as well as the
necesasry BOM or repository to add. This information should be useful
enough for any tool to add additional dependencies to an existing
project or create one from scratch.

Closes gh-140