Fix provided scope with Gradle build
Make sure to register the necessary configuration entry for Gradle when
a provided dependency is added as the build will fail otherwise.

Interestingly enough, this was only taken care of if the packaging is war
as we add the tomcat starter explicitly. We rather now add any provided
dependencies to the build with no special handling there, adding the
tomcat server in the model if necessary.

Closes gh-164