Dependency versions mapping support
If a dependency is fist managed by the service and later on by Spring
Boot, there is no way to keep a single entry that would change the
dependency according to the Spring Boot version.

A hack would consist of creating two entries that would have an exclusive
range but it's far from ideal.

This commit adds a `versions` attribute on each dependency that can
define an arbitrary number of mappings between a version and a range. If
one of those range matches, the related version is used (and the absence
of version means that no version need to be included). If no mapping
matched (or if there are no mappings at all), the default version is
used.

Closes gh-168