Upgrade to GMavenPlus
A groovy-based project using Maven now builds with GMavenPlus as the
Groovy Eclipse maven compiler does not seem to be maintained anymore.

This commit also harmonizes the location of the code to be
`src/main/groovy` for both Gradle and Maven builds.

Joint compilation is enabled by default so that you can create pure Java
code in `src/main/java` and refer to any class within the project.

Users of IntelliJ IDEA may want to install a plugin that automatically
adds `src/main/groovy` to the sources when the GMavenPlus plugin is
added to the pom, see https://plugins.jetbrains.com/plugin/7442

Closes gh-90