Use spring-cloud-dependencies
Previously, Spring Initializr was refering to
`spring-cloud-starter-parent` as the BOM source for any cloud-related
dependencies.

This pom actually inherits from a specific version of Spring Boot and
brings all its dependency management. If the chosen boot version and the
one that the cloud bom brings do not match, inconsistent versions can
appear in the project.

The underlying issue was fixed by providing a new BOM that only provides
dependency management for the cloud-related bits.

Closes gh-116