Fire event if a project could not be generated
Previously, only an invalid type or an invalid dependency would lead to
an exception and in such case, no event is fired at all.

This commit adds validation for language and packaging as well as a new
event that is fired when the project could not be generated.

The metrics infrastructure has been updated to handle ProjectFailedEvent;
when such an event is fired, the 'failures' counter is increased and we
still record all the other metrics.

Closes gh-188