Add clientId metrics
Parse the `User-Agent` http request header to determine if the client is
known. If that is the case, increment the relevant `client_id` counter.

A `ProjectRequest` had now a generic `parameters` array with all the
request headers by default. Since we don't want to accidentally map any
of those from a form input, `BasicProjectRequest` contains only "public"
fields and is the type exposed at the controller level.

Closes gh-193