Use the "name" property instead of the applicationName to set the spring.application.name
We guessed wrong in our first pass - the "name" property matches up to the value that the user passes in. The "ApplicationName" is not typically passed in, but is normally generated and is also used for the name of the main class file. Using the "name" property for the application name works better for us