Support for compileOnly scope
start.spring.io exposes an annotation processor (Lombok) that's quite
popular but is exposed as a compile/transitive dependency. This commit
introduces a `compileOnly` scope that maps to Gradle's `compileOnly`
scope and Maven's `optional` flag.

There is a case to support optional dependencies in general but
unfortunately, Gradle doesn't have support for such feature.

Closes gh-128