Fix typos
link → links

Links seems to be the correct key name according to the code and
semantically for a list, so all instances of `link:` have been replaced
by `links:`. Additionally, the Actuator reference link was pointing to
the wrong section in the documentation and has been fixed.

ref → rel

The application failed on startup due to `rel` being required.

Closes gh-360