Improve caching configuration
Rather than creating a `CacheManager` ourselves, we now rely on JCache
to provide a `CacheManager` that we'll augment with the caches that
initializr needs.

If Initializr is embedded in another app and that app uses JCache, then
its cache manager will be transparently upgraded to define the caches
that initializr needs. If not, such caches will have to be created as
part of the custom config.

Also, caching only kicks in if required at the service level.

This commit also switches the cache implementation to EhCache 3.

See gh-389