Add explicit support for Spring Cloud and Spring Boot 2
Spring Cloud does not currently work with Spring Boot 2.0.0.M4 and
later. This commit automatically downgrades to a compatible Spring Boot
version when a Spring Cloud artifact is requested.

It doesn't do that for SNAPSHOT to give a chance to Spring Cloud users
to test and validate the latest state of development.