Allow generation of custom files
Spring initializr allows to be customization so that it can be used in corporate environments or other pages than https://start.spring.io/.

To make it even more convenient to use generated projects right away is to generate additional files which are needed for CI tools e.g. Jenkins. In our case we would like to generate Jenkinsfiles which can be used as is.

One way to generate additional files is to override the method doGenerateProjectStructure:

```
    public class MyProjectGenerator extends ProjectGenerator {
        protected final Runnable jenkinsfileGenerated = Mockito.mock(Runnable.class);

        @Override
        protected File doGenerateProjectStructure(ProjectRequest request) {
            File dir = super.doGenerateProjectStructure(request);
            if (isMavenBuild(request)) {
                write(new File(dir, "Jenkinsfile"), "Jenkinsfile.tmpl", resolveModel(request));
                jenkinsfileGenerated.run();
            }
            return dir;
        }
    }
```

The problem here is that `doGenerateProjectStructure` publishes a ProjectGeneratedEvent. The issue is that with this setup ProjectGeneratedEvent would be published before the generation of files is actually done. Therefor this PR proses moving publishing of events to "generateProjectStructure".

To make it possible to reuse `isMavenBuild` and `isGradleBuild` their visibility has been changed to protected.