Create a settings.gradle file for Gradle projects
It would be beneficial to create a settings.gradle file for Gradle projects and add the "rootProject.name = '<artifact name>' to it. This is the default behavior when doing a gradle init. It also provides a specified project.name when doing builds and can be used by different Gradle plugins. Fixes gh-641.