Replace junit4 for junit5 when version >= 2.1.0.M4
Sam Brannen showed in his presentation (https://youtu.be/K7g2HUhWbNE) that for springboot 2.1.0 @ExtendWith is not needed anymore. I discovered that spring initializer creates a junit 4 test as a default to demonstrate that
the context loads.

Since the advent of junit 5 jupiter, I think we should opt to have make default test to use junit 5.

That inspired me to change the spring initializer. I've changed the following:
- maven pom, to include junit jupiter
- gradle build, to include junit jupiter
- java test, update test imports and remove @RunWith
- kotlin test, update test imports and remove @RunWith
- groovy test, update test imports and remove @RunWith

I didn't saw an test to test the context loads test, so I preformed it manually. 

Please let me know if I need to change something or need to add a test to test the generated testclass.

I thought that this change is trivial one and a ticket is not needed, however if I'm mistaken, please let me know