Remove default Web UI from Initializr library
Given that Web UIs can be implemented in many different ways and that
the start.spring.io UI is about to change dramatically.
Since Web UIs can be rendered on the server, or on the client side (or a
mix of both), provivind a default UI or tools to build a generally
useful one is not easy.

This commit removes all the Web UI specific code and leaves the Web
endpoints for REST access and metadata access.

Closes gh-854