Add support for lists with nested structure
Previously, a nested structure of a Maven Plugin configuration can be
only configured for a single parameter. This commit adds a dedicated
method to add a new entry with a nested structure every time it is
called. This effectively allows to configure nested "list" structure
such as listing dependency exclusions.

Closes gh-1165