Add support for Docker Compose
All DockerComposeService beans are added to the DockerComposeFile.
If that file is not empty, the DockerComposeProjectContributor will
write it as compose.yaml to the project.

The DockerComposeFileCustomizer serves as a hook-point to further
customize the DockerComposeFile.

The DockerComposeHelpDocumentCustomizer writes out a section into
the help file containing a table listing all the services with
their images and tags.