Introduce CodeBlock
This commit introduces a way to build more complex statement, including
multi statements and nested method invocations. CodeBlock is heavily
inspired from Square's JavaPoet project.

A CodeBlock is an immutable piece of code that provide the imports that
are needed and can emit code to an IndentingWriter.

This commit also migrates the use of expressions in Java, Kotlin, and
Groovy to the newly introduced CodeBlock. Those are deprecated in
favor of CodeBlock.

Closes gh-1043