Overhaul annotation support for code generation
This commit improves the annotation support so that it handles more
attribute types. An AnnotationContainer is introduced that permits to
further configure an annotation, in particular attribute values can be
added, and an attribute can be removed.

This commit adds a number of deprecations:

* Annotation#getName has been deprecated in favor of
Annotation#getClasName.
* Annotable#annotate has been deprecated in favor of
Annotable#annotations()#add.
* Annotable#getAnnotations has been deprecated in favor of
Annotable#annotations()#values.
* Annotation#name has been deprecated in favor of Annotation#of.
* Annotation.Builder#attribute has been deprecated in favor of
Annotation.Builder#set.

In particular, the type of the attribute does not need to be specified
as it is infered from the specified values. Use Annotation.Builder#add
or Annotation.Builder#set depending on your needs.

Closes gh-1424