Only accept GET or POST or both as approriate
All the endpoints in initializr use @RequestMapping with no explicit RequestMethod,
probably for historical reasons (@GetMapping and friends didn't exist when we
started this project). It would be more expressive to show which methods are expected
to be used, and I doubt anyone actually relies on being able to PATCH, PUT or POST
these endpoints. They might get a surprise if they tried to DELETE as well.